@@ -16,8 +16,8 @@
     "next": "^7.0.2",
     "nprogress": "^0.2.0",
     "prop-types": "^15.6.2",
-    "react": "^16.7.0",
-    "react-dom": "^16.7.0",
+    "react": "^16.8.0-alpha.1",
+    "react-dom": "^16.8.0-alpha.1",
     "styled-components": "^4.1.3"
   },
   "devDependencies": {
